+++ Samples/Image Processing/CropAndResizeRasterImage/Program.cs
@@ -10,10 +10,6 @@ namespace CropAndResizeRasterImage
     {
         static void Main()
         {
-
-            // Initialize evaluation license.
-            ImGearEvaluationManager.Initialize();
-
             // Initialize common formats.
             ImGearCommonFormats.Initialize();