+++ Samples/Image/LoadingAndSaving/Program.cs
@@ -9,10 +9,6 @@ namespace LoadingAndSaving
     {
         static void Main()
         {
-
-            // Initialize evaluation license.
-            ImGearEvaluationManager.Initialize();
-
             // Initialize common formats.
             ImGearCommonFormats.Initialize();