+++ Samples/Image/MetadataGeneralAPI/Program.cs
@@ -46,9 +46,6 @@ namespace MetadataGeneralAPI
 
         static void Main()
         {
-            // Initialize evaluation license.
-            ImGearEvaluationManager.Initialize();
-
             // Initialize common formats.
             ImGearCommonFormats.Initialize();