+++ Samples/PDF Acroform/CreatePDFFormField/Program.cs
@@ -11,10 +11,6 @@ namespace CreatePDFFormField
     {
         static void Main()
         {
-
-            // Initialize evaluation license.
-            ImGearEvaluationManager.Initialize();
-
             // Initialize common formats.
             ImGearCommonFormats.Initialize();
 
@@ -25,7 +21,7 @@ namespace CreatePDFFormField
             ImGearPDF.Initialize();
 
             // Load PDF document from a file.
-            using (Stream stream = new FileStream(@"../../../../../../Sample Input/accusoft-brochure.pdf", FileMode.Open, FileAccess.Read, FileShare.None))
+            using (Stream stream = new FileStream(@"../../../../../../../Sample Input/accusoft-brochure.pdf", FileMode.Open, FileAccess.Read, FileShare.None))
             {
                 using (ImGearPDFDocument pdfDocument = (ImGearPDFDocument)ImGearFileFormats.LoadDocument(stream))
                 {
@@ -64,7 +60,7 @@ namespace CreatePDFFormField
                     }
 
                     // Save the PDF document.
-                    using (Stream streamOut = new FileStream(@"../../../../../../Sample Output/CreatePDFFormField.pdf", FileMode.Create, FileAccess.Write, FileShare.None))
+                    using (Stream streamOut = new FileStream(@"../../../../../../../Sample Output/CreatePDFFormField.pdf", FileMode.Create, FileAccess.Write, FileShare.None))
                         ImGearFileFormats.SaveDocument(pdfDocument, streamOut, 0, ImGearSavingModes.OVERWRITE, ImGearSavingFormats.PDF, new ImGearSaveOptions());
                 }
             }