+++ Samples/PDF and JPEG/PlaceImageOnAnExistingPDFPage/Program.cs
@@ -11,10 +11,6 @@ namespace PlaceImageOnAnExistingPDFPage
     {
         static void Main()
         {
-
-            // Initialize evaluation license.
-            ImGearEvaluationManager.Initialize();
-
             // Initialize common formats.
             ImGearCommonFormats.Initialize();
 
@@ -25,12 +21,12 @@ namespace PlaceImageOnAnExistingPDFPage
             ImGearPDF.Initialize();
 
             // Load a PDF document.
-            using (Stream stream = new FileStream(@"../../../../../../Sample Input/multi-page.pdf", FileMode.Open, FileAccess.Read, FileShare.None))
+            using (Stream stream = new FileStream(@"../../../../../../../Sample Input/multi-page.pdf", FileMode.Open, FileAccess.Read, FileShare.None))
             {
                 using (ImGearPDFDocument pdfDocument = (ImGearPDFDocument)ImGearFileFormats.LoadDocument(stream))
                 {
                     // Load a raster image.
-                    using (Stream imageFileStream = new FileStream(@"../../../../../../Sample Input/water.jpg", FileMode.Open, FileAccess.Read))
+                    using (Stream imageFileStream = new FileStream(@"../../../../../../../Sample Input/water.jpg", FileMode.Open, FileAccess.Read))
                     {
                         ImGearPage loadedIgRasterPage = ImGearFileFormats.LoadPage(imageFileStream);
 
@@ -45,7 +41,7 @@ namespace PlaceImageOnAnExistingPDFPage
                     }
 
                     // Save the PDF document.
-                    using (Stream streamOut = new FileStream(@"../../../../../../Sample Output/PlaceImageOnAnExistingPDFPage.pdf", FileMode.Create, FileAccess.Write, FileShare.None))
+                    using (Stream streamOut = new FileStream(@"../../../../../../../Sample Output/PlaceImageOnAnExistingPDFPage.pdf", FileMode.Create, FileAccess.Write, FileShare.None))
                         ImGearFileFormats.SaveDocument(pdfDocument, streamOut, 0, ImGearSavingModes.OVERWRITE, ImGearSavingFormats.PDF, new ImGearSaveOptions());
                 }
             }