+++ Samples/PDF/PDFContentAddText/Program.cs
@@ -9,10 +9,6 @@ namespace PDFContentAddText
     {
         static void Main()
         {
-
-            // Initialize evaluation license.
-            ImGearEvaluationManager.Initialize();
-
             // Initialize common formats.
             ImGearCommonFormats.Initialize();
 
@@ -68,7 +64,7 @@ namespace PDFContentAddText
             textElement.Add(ImGearPDETextFlags.RUN, 0, textToAdd, font, gState, null, textMatrix, null);
 
             // Load a PDF document.
-            using (Stream stream = new FileStream(@"../../../../../../Sample Input/samplepdf.pdf", FileMode.Open, FileAccess.Read, FileShare.Read))
+            using (Stream stream = new FileStream(@"../../../../../../../Sample Input/samplepdf.pdf", FileMode.Open, FileAccess.Read, FileShare.Read))
             {
                 using (ImGearPDFDocument imGearPDFDocument = ImGearFileFormats.LoadDocument(stream) as ImGearPDFDocument)
                 {
@@ -88,7 +84,7 @@ namespace PDFContentAddText
                     }
 
                     // Save the PDF document.
-                    using (Stream outputStream = new FileStream(@"../../../../../../Sample Output/PDFContentAddText.pdf", FileMode.Create))
+                    using (Stream outputStream = new FileStream(@"../../../../../../../Sample Output/PDFContentAddText.pdf", FileMode.Create))
                         ImGearFileFormats.SaveDocument(imGearPDFDocument, outputStream, 1, ImGearSavingModes.OVERWRITE, ImGearSavingFormats.PDF, new ImGearSaveOptions());
                 }
             }