+++ Samples/PDF/SaveAllImagesOnPDFPage/Program.cs
@@ -48,7 +48,7 @@ namespace SaveAllImagesOnPDFPage
                         ImGearPage igPage = igPDEImage.ToImGearPage();
 
                         // Save image page.
-                        using (Stream outputStream = new FileStream(String.Format(@"../../../../../../Sample Output/SaveAllImagesOnPDFPage_{0}.png", imageIndex), FileMode.Create))
+                        using (Stream outputStream = new FileStream(String.Format(@"../../../../../../../Sample Output/SaveAllImagesOnPDFPage_{0}.png", imageIndex), FileMode.Create))
                             ImGearFileFormats.SavePage(igPage, outputStream, 1, ImGearSavingModes.OVERWRITE, ImGearSavingFormats.PNG);
                         imageIndex++;
                         break;
@@ -95,10 +95,6 @@ namespace SaveAllImagesOnPDFPage
 
         static void Main()
         {
-
-            // Initialize evaluation license.
-            ImGearEvaluationManager.Initialize();
-
             // Initialize common formats.
             ImGearCommonFormats.Initialize();
 
@@ -109,7 +105,7 @@ namespace SaveAllImagesOnPDFPage
             ImGearPDF.Initialize();
 
             // Load a PDF document.
-            using (Stream stream = new FileStream(@"../../../../../../Sample Input/samplepdf.pdf", FileMode.Open, FileAccess.Read, FileShare.Read))
+            using (Stream stream = new FileStream(@"../../../../../../../Sample Input/samplepdf.pdf", FileMode.Open, FileAccess.Read, FileShare.Read))
             {
                 using (ImGearPDFDocument igPDFDocument = (ImGearPDFDocument)ImGearFileFormats.LoadDocument(stream, 0, -1))
                 {